chore: rename `pool` -> `aio_helper`
